Fix modal wrapper close on submit status form
